added exceptions + changes to service

+++ src/main/java/com/codedifferently/groupone/SpyGlass/services/GoalService.java
@@ -1,6 +1,7 @@
 package com.codedifferently.groupone.SpyGlass.services;
 
 import com.codedifferently.groupone.SpyGlass.entities.Goal;
+import com.codedifferently.groupone.SpyGlass.exceptions.goal.GoalNotFoundException;
 import com.codedifferently.groupone.SpyGlass.repos.GoalRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
@@ -18,20 +19,45 @@ public class GoalService {
     @Autowired
     private GoalRepo goalRepo;
 
+    /**
+     * @return list of all goals
+     */
     public List<Goal> getGoals() {
         return goalRepo.findAll();
     }
 
-    public Goal getGoalById(@PathVariable Long id) {
-        return goalRepo.findById(id).orElseThrow(RuntimeException::new);
+    /**
+     * @param id
+     * @return goal found by id
+     * @throws GoalNotFoundException
+     */
+    public Goal getGoalById(@PathVariable Long id) throws GoalNotFoundException {
+        if (!goalRepo.existsById(id)) {
+            throw new GoalNotFoundException(id);
+        }
+        return goalRepo.getById(id);
     }
 
+    /**
+     * creates new goal
+     * @param goal
+     * @throws URISyntaxException
+     */
     public ResponseEntity addGoal(@RequestBody Goal goal) throws URISyntaxException {
         Goal newGoal = goalRepo.save(goal);
         return ResponseEntity.created(new URI("/goals" + newGoal.getId())).body(newGoal);
     }
 
-    public ResponseEntity editGoal(@PathVariable Long id, @RequestBody Goal goal) {
+    /**
+     * edits goal
+     * @param id
+     * @param goal
+     * @throws GoalNotFoundException
+     */
+    public ResponseEntity editGoal(@PathVariable Long id, @RequestBody Goal goal) throws GoalNotFoundException {
+        if (!goalRepo.existsById(id)) {
+            throw new GoalNotFoundException(id);
+        }
         Goal editingGoal = goalRepo.findById(id).orElseThrow(RuntimeException::new);
         editingGoal.setGoalAmount(goal.getGoalAmount());
         editingGoal.setContributionAmount(goal.getContributionAmount());
@@ -42,13 +68,18 @@ public class GoalService {
         return ResponseEntity.ok(editingGoal);
     }
 
-    public ResponseEntity deleteGoal(@PathVariable Long id) {
+    /**
+     * deletes goal
+     * @param id
+     * @throws GoalNotFoundException
+     */
+    public ResponseEntity deleteGoal(@PathVariable Long id) throws GoalNotFoundException {
+        if (!goalRepo.existsById(id)) {
+            throw new GoalNotFoundException(id);
+        }
         goalRepo.deleteById(id);
         return ResponseEntity.ok().build();
     }
-
-
-
 
 
 }
